Allow yielding through cqueue_resume => cqueue_process => cqueues_step.
Had to add a "current thread" out-arg to cqueue_process to know which thread's stack to push resumed values back onto.
Seemed a disappointment to not be able to re-use the 'error.thread' field of the callinfo structure.

I had to duplicate a large amount of the cqueue_step function as a continuation;
the continuation and the original cqueue_step function should always be kept in sync.