Add workaround for Lua 5.1.
You can't pass a continuation in C; but you can in Lua!
This adds a 'set_resume' method that only appears in lua 5.1:
it un-resumes the currently yielded thread in a cqueue, but doesn't run anything yet.
We then wrap 'step' in 5.1 to check for yields, and send them via 'set_resume' before stepping

The way I coded this, a given coroutine may run twice in a single step in Lua 5.1 if another gets yielded,
and it has a short timeout that elapses before the yielded thread is resumed.